rationalising unit tests
